@@ -3,7 +3,7 @@
   "description": "rabbit httpserver",
   "type": "library",
   "require": {
-    "php": ">=7.4.0",
+    "php": ">=8.0.0",
     "ext-swoole": ">=4.5.0",
     "rabbit/base": "dev-master",
     "rabbit/web": "dev-master",
