@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Rabbit\HttpServer\Formater;
@@ -11,11 +12,7 @@ use Psr\Http\Message\ResponseInterface;
  */
 class ResponseRawFormater implements ResponseFormaterInterface
 {
-    /**
-     * @param ResponseInterface $response
-     * @return ResponseInterface
-     */
-    public function format(ResponseInterface $response, $data): ResponseInterface
+    public function format(ResponseInterface $response, string|array|object|float|int|bool|null &$data): ResponseInterface
     {
         // Headers
         $response = $response->withoutHeader('Content-Type')->withAddedHeader('Content-Type', 'text/plain');
