@@ -15,67 +15,29 @@ use Psr\Http\Message\ServerRequestInterface;
 class Request implements ServerRequestInterface
 {
     use MessageTrait;
-    /**
-     * @var \Swoole\Http\Request
-     */
+
     protected \Swoole\Http\Request $swooleRequest;
 
-    /**
-     * @var array
-     */
     private array $attributes = [];
 
-    /**
-     * @var array
-     */
     private array $cookieParams = [];
 
-    /**
-     * @var null|array|object
-     */
     private $parsedBody;
 
-    /**
-     * the body of parser
-     *
-     * @var mixed
-     */
     private $bodyParams;
 
-    /**
-     * @var array
-     */
     private array $queryParams = [];
 
-    /**
-     * @var array
-     */
     private array $serverParams = [];
 
-    /**
-     * @var array
-     */
     private array $uploadedFiles = [];
 
-    /**
-     * @var string
-     */
     private $method;
 
-    /**
-     * @var UriInterface
-     */
     private UriInterface $uri;
 
-    /**
-     * @var string
-     */
     private string $requestTarget;
 
-    /**
-     * Request constructor.
-     * @param \Swoole\Http\Request $swooleRequest
-     */
     public function __construct(\Swoole\Http\Request $swooleRequest)
     {
         $server = $swooleRequest->server;
@@ -93,10 +55,6 @@ class Request implements ServerRequestInterface
             ->setSwooleRequest($swooleRequest);
     }
 
-    /**
-     * @param array $headers
-     * @return $this
-     */
     private function setHeaders(array $headers): Request
     {
         $this->headers = [];
@@ -112,10 +70,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param \Swoole\Http\Request $swooleRequest
-     * @return Uri
-     */
     private static function getUriFromGlobals(\Swoole\Http\Request $swooleRequest): Uri
     {
         $server = $swooleRequest->server;
@@ -171,10 +125,6 @@ class Request implements ServerRequestInterface
         return $uri;
     }
 
-    /**
-     * @param array $serverParams
-     * @return Request
-     */
     public function withServerParams(array $serverParams): Request
     {
 
@@ -182,10 +132,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param array $uploadedFiles
-     * @return Request|static
-     */
     public function withUploadedFiles(array $uploadedFiles)
     {
 
@@ -193,10 +139,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param array|null|object $data
-     * @return Request|static
-     */
     public function withParsedBody($data)
     {
 
@@ -204,10 +146,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param array $query
-     * @return Request|static
-     */
     public function withQueryParams(array $query)
     {
 
@@ -215,10 +153,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param array $cookies
-     * @return Request|static
-     */
     public function withCookieParams(array $cookies)
     {
 
@@ -226,10 +160,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param array $files
-     * @return array
-     */
     private static function normalizeFiles(array $files): array
     {
         $normalized = [];
@@ -250,10 +180,6 @@ class Request implements ServerRequestInterface
         return $normalized;
     }
 
-    /**
-     * @param array $value
-     * @return array|UploadedFile
-     */
     private static function createUploadedFileFromSpec(array $value): UploadedFile
     {
         if (is_array($value['tmp_name'])) {
@@ -263,10 +189,6 @@ class Request implements ServerRequestInterface
         return new UploadedFile($value['tmp_name'], (int)$value['size'], (int)$value['error'], $value['name'], $value['type']);
     }
 
-    /**
-     * @param array $files
-     * @return array
-     */
     private static function normalizeNestedFileSpec(array $files = []): array
     {
         $normalizedFiles = [];
@@ -285,69 +207,41 @@ class Request implements ServerRequestInterface
         return $normalizedFiles;
     }
 
-    /**
-     * @return array
-     */
     public function getServerParams()
     {
         return $this->serverParams;
     }
 
-    /**
-     * @return array
-     */
     public function getCookieParams()
     {
         return $this->cookieParams;
     }
 
-    /**
-     * @return array
-     */
     public function getQueryParams()
     {
         return $this->queryParams;
     }
 
-    /**
-     * @return array
-     */
     public function getUploadedFiles()
     {
         return $this->uploadedFiles;
     }
 
-    /**
-     * @return array|null|object
-     */
     public function getParsedBody()
     {
         return $this->parsedBody;
     }
 
-    /**
-     * @return array
-     */
     public function getAttributes()
     {
         return $this->attributes;
     }
 
-    /**
-     * @param string $name
-     * @param null $default
-     * @return mixed|null
-     */
     public function getAttribute($name, $default = null)
     {
         return array_key_exists($name, $this->attributes) ? $this->attributes[$name] : $default;
     }
 
-    /**
-     * @param string $name
-     * @param mixed $value
-     * @return Request|static
-     */
     public function withAttribute($name, $value)
     {
 
@@ -355,10 +249,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @param string $name
-     * @return $this|Request|static
-     */
     public function withoutAttribute($name)
     {
         if (false === array_key_exists($name, $this->attributes)) {
@@ -371,9 +261,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getRequestTarget()
     {
         if ($this->requestTarget !== null) {
@@ -391,10 +278,6 @@ class Request implements ServerRequestInterface
         return $target;
     }
 
-    /**
-     * @param mixed $requestTarget
-     * @return Request|static
-     */
     public function withRequestTarget($requestTarget)
     {
         if (preg_match('#\s#', $requestTarget)) {
@@ -406,18 +289,11 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getMethod()
     {
         return $this->method;
     }
 
-    /**
-     * @param string $method
-     * @return Request|static
-     */
     public function withMethod($method)
     {
         $method = strtoupper($method);
@@ -430,19 +306,11 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     * @return UriInterface|Request|Uri
-     */
     public function getUri()
     {
         return $this->uri;
     }
 
-    /**
-     * @param UriInterface $uri
-     * @param bool $preserveHost
-     * @return $this|Request|static
-     */
     public function withUri(UriInterface $uri, $preserveHost = false)
     {
         if ($uri === $this->uri) {
@@ -459,9 +327,6 @@ class Request implements ServerRequestInterface
         return $this;
     }
 
-    /**
-     *
-     */
     private function updateHostFromUri(): void
     {
         $host = $this->uri->getHost();
@@ -483,18 +348,11 @@ class Request implements ServerRequestInterface
         $this->headers = [$header => [$host]] + $this->headers;
     }
 
-    /**
-     * @return \Swoole\Http\Request
-     */
     public function getSwooleRequest(): \Swoole\Http\Request
     {
         return $this->swooleRequest;
     }
 
-    /**
-     * @param \Swoole\Http\Request $swooleRequest
-     * @return $this
-     */
     public function setSwooleRequest(\Swoole\Http\Request $swooleRequest): Request
     {
         $this->swooleRequest = $swooleRequest;
