@@ -16,15 +16,11 @@ use Rabbit\Base\Exception\NotSupportedException;
 class Response implements ResponseInterface
 {
     use MessageTrait;
-    /**
-     * @var int
-     */
+
     private int $statusCode = 200;
-    /**
-     * @var string
-     */
+
     private string $charset = 'utf-8';
-    /** @var \Swoole\Http\Response */
+
     protected \Swoole\Http\Response $swooleResponse;
 
     protected array $fdList = [];
@@ -35,37 +31,22 @@ class Response implements ResponseInterface
         return $this;
     }
 
-    /**
-     * @return int
-     */
     public function getStatusCode()
     {
         return $this->statusCode;
     }
 
-    /**
-     * @param int $code
-     * @param string $reasonPhrase
-     * @return mixed|static
-     */
     public function withStatus($code, $reasonPhrase = '')
     {
         $this->statusCode = (int)$code;
         return $this;
     }
 
-    /**
-     * @return string|void
-     * @throws NotSupportedException
-     */
     public function getReasonPhrase()
     {
         throw new NotSupportedException("can not call " . __METHOD__);
     }
 
-    /**
-     *
-     */
     public function send(): void
     {
         foreach ($this->fdList as $fd => $message) {
@@ -78,28 +59,16 @@ class Response implements ResponseInterface
         });
     }
 
-    /**
-     * @param int $fd
-     * @param string $msg
-     * @throws Exception
-     */
     public function push(int $fd, string $msg): void
     {
         (new \Swoole\Http\Response($fd))->push($msg);
     }
 
-    /**
-     * @return \Swoole\Http\Response
-     */
     public function getSwooleResponse(): \Swoole\Http\Response
     {
         return $this->swooleResponse;
     }
 
-    /**
-     * @param \Swoole\Http\Response $response
-     * @return Response
-     */
     public function setSwooleResponse(\Swoole\Http\Response $response): self
     {
         $this->swooleResponse = $response;
